+++ pkgs/html/lib/dom_parsing.dart
@@ -81,23 +81,27 @@ class CodeMarkupVisitor extends TreeVisitor {
 
   @override
   void visitDocumentType(DocumentType node) {
-    _str.write('<code class="markup doctype">&lt;!DOCTYPE ${node.name}>'
+    var name = node.name ?? '';
+    if (name.isNotEmpty) name = ' ${htmlSerializeEscape(name)}';
+    _str.write('<code class="markup doctype">&lt;!DOCTYPE$name>'
         '</code>');
   }
 
   @override
   void visitText(Text node) {
-    writeTextNodeAsHtml(_str, node);
+    _str.write(htmlSerializeEscape(node.data));
   }
 
   @override
   void visitElement(Element node) {
     final tag = node.localName;
-    _str.write('&lt;<code class="markup element-name">$tag</code>');
+    final escapedTag = tag != null ? htmlSerializeEscape(tag) : null;
+    _str.write('&lt;<code class="markup element-name">$escapedTag</code>');
     if (node.attributes.isNotEmpty) {
       node.attributes.forEach((key, v) {
-        v = htmlSerializeEscape(v, attributeMode: true);
-        _str.write(' <code class="markup attribute-name">$key</code>'
+        final escapedKey = htmlSerializeEscape(key.toString());
+        v = htmlSerializeEscape(v);
+        _str.write(' <code class="markup attribute-name">$escapedKey</code>'
             '=<code class="markup attribute-value">"$v"</code>');
       });
     }
@@ -107,8 +111,10 @@ class CodeMarkupVisitor extends TreeVisitor {
     } else if (isVoidElement(tag)) {
       _str.write('>');
       return;
+    } else {
+      _str.write('>');
     }
-    _str.write('&lt;/<code class="markup element-name">$tag</code>>');
+    _str.write('&lt;/<code class="markup element-name">$escapedTag</code>>');
   }
 
   @override